+++ packages/terra-clinical-header/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Changed
+  * Replaced `title` prop to `text` prop to allow users to spread title HTML Attribute as customProps.
+  * Added `level` prop to allow users to render different heading levels.
+  * Added `id` prop which is to be paired with `aria-labelledby` for labelling page regions.
+
 ## 3.26.0 - (April 21, 2023)
 
 * Changed