+++ packages/terra-clinical-onset-picker/CHANGELOG.md
@@ -15,6 +15,9 @@
   * Updated wdio screenshots.
   * Updated jest snapshots
   * Locking terra-form-fieldset due to a bug
+* Changed
+  * Updated jest snapshots terra-icon A11y updatess.
+
 
 ## 4.31.1 - (April 19, 2022)